Extracts test utilities for native number index testing
